backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
pymupdf>=1.25.0
anthropic>=0.40.0
python-dotenv>=1.0.0
gspread>=6.0.0
